requirements.txt
asyncio==3.4.3
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cloudevents==1.9.0
deprecation==2.1.0
dnspython==2.3.0
Flask==2.2.3
functions-framework==3.3.0
geojson==3.0.1
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==1.0.0
google-cloud-appengine-logging==1.3.0
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.9.0
google-cloud-core==2.3.2
google-cloud-logging==3.5.0
google-cloud-pubsub==2.16.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.53.0
grpcio-status==1.53.0
gspread==5.7.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.3.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==2.1.1
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.0
shapely==2.0.1
six==1.16.0
tenacity==8.2.2
tqdm==4.64.1
turfpy==0.0.7
tzdata==2023.3
urllib3==1.26.14
watchdog==3.0.0
Werkzeug==2.2.3
zipp==3.15.0
